Add a basic command execution feature
Whenever the input contains a newline character (anywhere in the string), we
strip all newlines and begin to execute the command. Currently it just supports
"next" and "prev" for now.

Also, make greater use of instaparse's hide facility to produce a cleaner parse
tree. Use `:unhide :all` to continue to produce a full parse for display in the
DOM with syntax highlighting.