Add state restoration using `window.localStorage`
Also a Closure define of VERBOSE is now here to control whether we should
print. (It appears that `enable-console-print!` is already enabled even in
release builds.)

The reload function just catches all exceptions, not just from localStorage
interaction, since it is observed that a sufficiently corrupt localStorage can
cause `js->clj` to fail.

The app-state atom has been split to capture the session state that does not
need to be stored.