Implement the complete add command
The complicated recurrence specification can be parsed, understood, and
transformed into the right structures now.

Also refactored the built-in bank holidays events to be based on the new
recurrent event system.

Also changed the event storage from a map to a simple flat vector. We can always
optimize that further.