Add an `ls` command to list all events alphabetically
It shows, for each event, the name, the recurring pattern, the occurrences, and
a minigrid of occurrences inspired by sparklines.

(The minigrid currently may be too long. Will probably restrict its range
somewhat.)

The command can show either all events or only those that are visible. This is
explained in the help.