Introduce functions
This implementation of functions is entirely different from larger languages in
that it is entirely parser directed. Instead of a parser simply parsing it as a
function-call syntax with expressions, and then later going through name
resolution and type checking, everything is built into the parser. This means
incorrectly typed functions are parse errors. It also means if we choose to
have function overloading, it will be directed by return type rather than
argument types. But then again, we will need to duplicate the syntactic rules
for function calls across each function (but we can switch to parser combinators
if needed.)

Since it is directed by the parser, `plus(Dec 1, 2020, 1d)` is valid syntax even
though the function takes two arguments separated by a comma: the first comma is
part of the date literal.

Also introduce a new date interval type. They may be constructed via functions
or literals. They can only be used in the second argument of `plus` and `minus`
functions. The beauty is that we don't even track this type explicitly.

Also fix some whitespace specification so that `every 2 days` and `every(2)days`
are both acceptable.