Add mouse-based interactions to the modal
When the user clicks an event, the modal is shown to display details of that
event. When the user clicks on the calendar, the modal is dismissed.

Furthermore, the modal can now be shorter than 85vh.

Furthermore, when the modal is hidden, the content is still in the DOM to allow
the modal to display the animation with contents.