Rewrite the UI components in Shadow Grove rather than Reagent
After reading its documentation (which as the author knows, is incomplete),
I became somewhat enamored of its design and its commitment to performance. This
is also an opportunity to improve performance in my own code.

I also took this as an opportunity to improve some aspects: the
textarea-change (which is technically `onInput` rather than `onChange` in React
land) event logic is simplified by simply removing the prompt as a subsequence.

An interesting consequence of Grove not doing VDOM diffing is that we cannot
have a conditional that produces the same element in all branches. This will
lead to Grove deleting and recreating the element. This would have been
unnoticeable except for the fact that it breaks CSS transitions.

This change reduces the final output JS file by about 100KB.