History search in the command-line, inspired by the fish shell
The state being kept here is surprisingly intricate.

If the user does not interact with the feature, the current completion candidate
is displayed. Otherwise the user interacts with the feature using the four arrow
keys. The right arrow key commits the current history-based completion. The up
and down arrow keys navigate through the history. The left arrow key behaves
normally but does not exit the history navigation.

Ctrl-R is purposefully not used because it clashes with the browser refresh
feature and would be undesirable.

The code here is somewhat long and messy but I hope to refactor this soon. The
logic is subtle but by Clojure standards this is too much code.