Refactor the representation of history state
Instead of a nested map, it becomes a flattened single-layer map. This makes the
manipulation code simpler.

Interestingly, Grove complains that assoc is not allowed on observed data:
https://github.com/thheller/shadow-grove/blob/bc580434fb208f5de33af6e71fb1cfeb8d48f9b5/src/main/shadow/grove/kv.cljs#L30
so I got rid of the assoc. It is not required anyways. Doing so is probably good
for performance too.

Speaking of performance, Grove may be able to deduce fewer re-renders due to the
more granular access pattern being known to Grove.