Prevent memory leak caused by subvec not releasing memory
The problem itself manifests clearly when we look at the following REPL session
using subvec:

```
(subvec [0 1 2 3 4] 2)
[2 3 4]

(conj (subvec [0 1 2 3 4] 2) 5)
[2 3 4 5]

(.-v (conj (subvec [0 1 2 3 4] 2) 5))
[0 1 2 3 4 5]

(.-v (conj (subvec [0 1 2 3 4] 2 3) 5))
[0 1 2 5 4]
```

Indeed, subvec is a lightweight wrapper over a regular vector. When we conj onto
it, it simply produces a new vector with the correct index reset to the given
value. It never releases memory for the front.

This is solved by periodically calling `(into [])` on the subvec.

This slightly complicates the history search feature by making it ignore entries
that should have been deleted.