Support multiple occurrences/recurrences for an event
This is quite a fundamental change to the data model. An event can now be
associated with multiple occurrences and multiple recurrence rules. These
multiple occurrences and recurrence rules are implicitly unioned together. This
is essential to supporting features such as moving a single occurrence in a
recurrent series.

I hope to introduce some recurrence rule optimization in the future when we
merge events.