Adjust the recurrence start date to the first occurrence
This was originally envisioned as a helper function to the event
canonicalization work, but doing so seems useful in its own right and provides a
better user experience to show the actual next occurrence in the ls modal.
Furthermore, if the event is determined not to occur, show a helpful error
message.

The first occurrence is simply calculated by repeated iteration with an upper
bound. The upper bound is chosen to be as small as possible while remaining
correct. In some cases, additional analysis is done (e.g. whether the monthly
recurrence is only on the 29th and February is involved) to determine a tighter
upper bound. This is helpful because we wish to return that error message as
soon as possible. Later on, we may wish to use jsbench or some such to do some
tests to confirm this is worthwhile.