Set :fn-invoke-direct
This option calls higher-order functions simply using parenthesis, rather than
going through Function.prototype.call. The difference is that the `this`
argument becomes whatever the object the function is called on (probably a
Closure module) rather than the original. This should be fixed in the original
code by using Function.prototype.bind anyways.