Change the semantics of week-based recurrences
Originally this simply selects the soonest days-of-week matching the spec and
sets the recurrent period to be 7 times the number of weeks.

The new behavior more closely matches the month and year based recurrences and
is more intuitive. It first selects the weeks and then within the week selects
the relevant days of week.

The difference is only apparent when the start date is past the first selected
day-of-week in the current week. Unit tests are modified.

Also take the opportunity to reduce code duplication in the
recurrent-event-occurrences function.