Refactor the select-dates-from-*-recur fns to take multiple periods
These usually need to do a bit of preprocessing such as sorting that can be
shared so that it does not need to be repeated for each period. Furthermore, the
find-first-start function is simplified to use `lazy-seq` and `cons` rather than
writing out the same computation twice connected by `or`.