Add an event rule optimization
This tries to optimize the recurrence rules in an event to be more easily
comprehensible. Exactly what is more easily comprehensible is subjective, so this
is just currently whatever proposed in the unit tests. I'm sure there will be
future adjustments to this optimization logic.

The optimization logic is yet to be implemented for week/month/year recurrences.

The coding style has shifted by a bit; there are more uses of transients and the
loop construct. However in my current style I do not allow transients to escape
a single function. The loop construct is in fact more convenient than other
languages with guaranteed tail-call optimization; but still unrestricted
recursion is like the goto for functional programming so I still make an effort
to avoid it.