Add more configuration options
The first is to enable or disable event optimization. I think some users simply
prefer not to have the recurrence changes changed on their back.

The second is to disable autocomplete. Some users find that it destroys their
train of thought, cf https://hachyderm.io/@robpike/113541673495454275. It's the
same reason why I disable autocompletion in my own editors and my browser search
bar.

The third is to disable explanation of commands. This can cause shifting of
layouts (when a one-line command becomes two lines) and may be distracting to
motion sensitive users.