Format single-occ recurrences in a nicer way
These single-occ recurrences are often the result of splitting recurrences
during event optimization. It seems like a nice idea to optimize the formatting
of these recurrences.

For week-based recurrences, the displayed format is simplified into a single day
if possible, or days within a single week. It is anticipated that something
similar will be done for month- and year-based recurrences.