Refine the condition for front and back partial weeks
Before this commit, a condition is to require a recurrence not to start on a week
boundary. After this commit, a condition is to require there is at least one
occurrence in the first period before the recur-start. This preserves the
freq and remainder in more cases.

This is a bit slower than the old code path, but is more amenable to
generalization to other recur-types.