Eliminate the double parsing of the string
Currently, besides parsing the string for syntax highlighting purposes, we parse
the string twice more, for explaining the input and for actual execution. The
latter two can be consolidated by introducing an extra binding in the component.

This consolidated parse can be skipped if the earlier parse for syntax
highlighting already results in a failure.

Also, remove the use of throwing exceptions for failure: we don't actually deal
with exceptions from `transform-parsed` very well. We probably will want to
switch to a non-exception based error library.